Do not download plugins that haven't had an update - done via adding versions to files.

+++ providers.py
@@ -1,8 +1,14 @@
 import requests
 import hash
 
+import os
 
-def download(plugin, server_ver, headers, install):
+
+def plugin_version_exists(jarfile):
+    return os.path.exists("./plugins/" + jarfile)
+
+
+def download(plugin, headers, install):
     name = plugin["name"]
     provider = plugin["provider"]
     plugin_id = plugin["id"]
@@ -13,32 +19,27 @@ def download(plugin, server_ver, headers, install):
 
     if provider == "modrinth":
         print("Downloading " + name + " from Modrinth...")
-        response = requests.get('https://api.modrinth.com/v2/project/' + plugin_id, headers=headers,
-                                allow_redirects=True)
-        response = response.json()
-        if server_ver in response["game_versions"]:
-            params = {
-                'game_versions': '["' + server_ver + '"]',
-                'loaders': '["paper", "spigot", "bukkit"]'
-            }
-            versions = requests.get('https://api.modrinth.com/v2/project/' + plugin_id + "/version", params=params,
-                                    headers=headers, allow_redirects=True)
-            versions = versions.json()
-            expected_hash = versions[0]["files"][0]["hashes"]["sha512"]
-            url = versions[0]["files"][0]["url"]
-            filename = versions[0]["files"][0]["filename"]
-            filename = path + filename
+        params = {
+            'loaders': '["paper", "spigot", "bukkit"]'
+        }
+        versions = requests.get('https://api.modrinth.com/v2/project/' + plugin_id + "/version", params=params,
+                                headers=headers, allow_redirects=True)
+        versions = versions.json()
+        expected_hash = versions[0]["files"][0]["hashes"]["sha512"]
+        url = versions[0]["files"][0]["url"]
+        filename = versions[0]["files"][0]["filename"]
+        if plugin_version_exists(filename):
+            print("No updates")
+            return
+        filename = path + filename
+        response = requests.get(url)
+        with open(filename, 'wb') as f:
+            f.write(response.content)
 
-            response = requests.get(url)
-            with open(filename, 'wb') as f:
-                f.write(response.content)
+        new_hash = hash.hasher(filename, 512)
 
-            new_hash = hash.hasher(filename, 512)
-
-            if not new_hash == str(expected_hash):
-                print("Bad hash, incorrect or unexpected file. Please Manually check the file.")
-        else:
-            print("There is no available version of " + name + " for " + server_ver + ". Sorry.")
+        if not new_hash == str(expected_hash):
+            print("Bad hash, incorrect or unexpected file. Please Manually check the file.")
 
     elif provider == "github":
         print("Downloading " + name + " from Github...")
@@ -46,7 +47,12 @@ def download(plugin, server_ver, headers, install):
                                 allow_redirects=True)
         releases = response.json()
         release = releases[0]  # Is a dictionary
-        filename = release["assets"][0]["name"]
+        newest_version = release["name"]
+        print(newest_version)
+        filename = newest_version + "-" + release["assets"][0]["name"]
+        if plugin_version_exists(filename):
+            print("No updates")
+            return
         filename = path + filename
         download_url = release["assets"][0]["browser_download_url"]
 
@@ -55,22 +61,28 @@ def download(plugin, server_ver, headers, install):
             f.write(new_file.content)
     elif provider == "spigot":
         print("Downloading " + name + " from Spigot...")
-        filename = path + name + ".jar"
+        date = requests.get("https://api.spiget.org/v2/resources/" + plugin_id, headers=headers,
+                            allow_redirects=True).json()["updateDate"]
+        filename = name + "-" + str(date) + ".jar"
+        if plugin_version_exists(filename):
+            print("No updates")
+            return
+        filename = path + filename
         response = requests.get("https://api.spiget.org/v2/resources/" + plugin_id + "/download", headers=headers,
                                 allow_redirects=True)
         with open(filename, 'wb') as f:
             f.write(response.content)
     elif provider == "geyser":
         print("Downloading " + name + " from the GeyserMC website...")
-        filename = path + name + "-Spigot.jar"
-        if plugin_id == "geyser":
-            response = requests.get(
-                "https://download.geysermc.org/v2/projects/geyser/versions/latest/builds/latest/downloads/spigot",
-                headers=headers, allow_redirects=True)
-        if plugin_id == "floodgate":
-            response = requests.get(
-                "https://download.geysermc.org/v2/projects/floodgate/versions/latest/builds/latest/downloads/spigot",
-                headers=headers, allow_redirects=True)
+        version = requests.get("https://download.geysermc.org/v2/projects/" + plugin_id, headers=headers, allow_redirects=True).json()["versions"][-1]
+        filename = name + "-" + version + ".jar"
+        if plugin_version_exists(filename):
+            print("No updates")
+            return
+        filename = path + filename
+        response = requests.get(
+            "https://download.geysermc.org/v2/projects/" + plugin_id + "/versions/latest/builds/latest/downloads/spigot",
+            headers=headers, allow_redirects=True)
         with open(filename, 'wb') as f:
             f.write(response.content)
     elif provider == "hangar":
@@ -81,6 +93,9 @@ def download(plugin, server_ver, headers, install):
 
         url = response["result"][0]["downloads"]["PAPER"]["downloadUrl"]
         filename = response["result"][0]["downloads"]["PAPER"]["fileInfo"]["name"]
+        if plugin_version_exists(filename):
+            print("No updates")
+            return
         filename = path + filename
         response = requests.get(url)
         with open(filename, 'wb') as f:
